[ui]: move graduation component to parent folder (#9)
