+++ src/components/graduation/CreditInfoCard.tsx
