+++ src/components/graduation/GraduationStatus.tsx
