+++ src/components/graduation/GraduationTable.tsx
