+++ src/components/graduation/ResponsiveTable.tsx
